+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/bomtool/cran/PackRatNodeParser.groovy
@@ -29,21 +29,21 @@ import com.blackducksoftware.integration.hub.bdio.simple.model.DependencyNode
 import com.blackducksoftware.integration.hub.detect.Application
 import com.blackducksoftware.integration.hub.detect.bomtool.CranBomTool
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNode
-import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeBuilder
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeImpl
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeTransformer
+import com.blackducksoftware.integration.hub.detect.nameversion.builder.NameVersionNodeBuilderImpl
 
 public class PackRatNodeParser {
     private final Logger logger = LoggerFactory.getLogger(PackRatNodeParser.class)
 
     private NameVersionNode rootNameVersionNode
-    private NameVersionNodeBuilder nameVersionNodeBuilder
+    private NameVersionNodeBuilderImpl nameVersionNodeBuilder
     private HashSet<String> directDependencyNames
     private NameVersionNode currentParent
 
     List<DependencyNode> parseProjectDependencies(NameVersionNodeTransformer nameVersionNodeTransformer, final String packratLockContents) {
         rootNameVersionNode = new NameVersionNodeImpl([name: 'packratLockFileRoot'])
-        nameVersionNodeBuilder = new NameVersionNodeBuilder(rootNameVersionNode)
+        nameVersionNodeBuilder = new NameVersionNodeBuilderImpl(rootNameVersionNode)
         directDependencyNames = new HashSet<>()
         currentParent = null
