+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/bomtool/pip/PipInspectorTreeParser.groovy
@@ -30,9 +30,9 @@ import com.blackducksoftware.integration.hub.bdio.simple.model.DependencyNode
 import com.blackducksoftware.integration.hub.bdio.simple.model.Forge
 import com.blackducksoftware.integration.hub.detect.Application
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNode
-import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeBuilder
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeImpl
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeTransformer
+import com.blackducksoftware.integration.hub.detect.nameversion.builder.NameVersionNodeBuilderImpl
 
 @Component
 class PipInspectorTreeParser {
@@ -72,7 +72,7 @@ class PipInspectorTreeParser {
             if (line.contains(SEPARATOR) && !nodeBuilder) {
                 NameVersionNode projectNode = lineToNode(line)
                 tree.push(projectNode)
-                nodeBuilder = new NameVersionNodeBuilder(projectNode)
+                nodeBuilder = new NameVersionNodeBuilderImpl(projectNode)
                 continue
             }
