+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/nameversion/NameVersionNode.groovy
@@ -26,10 +26,10 @@ interface NameVersionNode {
     public String getName()
     public String getVersion()
     public List<? extends NameVersionNode> getChildren()
-    public Metadata getMetadata()
+    public NodeMetadata getMetadata()
 
     public void setName(String name)
     public void setVersion(String version)
     public void setChildren(List<? extends NameVersionNode> children)
-    public void setMetadata(Metadata metadata)
+    public void setMetadata(NodeMetadata metadata)
 }
