+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/nameversion/NameVersionNodeTransformer.groovy
@@ -30,6 +30,10 @@ import com.blackducksoftware.integration.hub.bdio.simple.model.externalid.NameVe
 
 @Component
 class NameVersionNodeTransformer {
+    public DependencyNode createDependencyNode(Forge defaultForge, NameVersionNodeBuilder nameVersionNodeBuilder) {
+        createDependencyNode(nameVersionNodeBuilder.build())
+    }
+
     public DependencyNode createDependencyNode(Forge defaultForge, NameVersionNode nameVersionNode) {
         final Forge forge = nameVersionNode.metadata?.forge ? nameVersionNode.metadata.forge : defaultForge
         def externalId = new NameVersionExternalId(forge, nameVersionNode.name, nameVersionNode.version)
