+++ src/main/groovy/com/blackducksoftware/integration/hub/detect/nameversion/NodeMetadata.groovy
@@ -1,0 +1,20 @@
+/*
+ * Copyright (C) 2017 Black Duck Software Inc.
+ * http://www.blackducksoftware.com/
+ * All rights reserved.
+ *
+ * This software is the confidential and proprietary information of
+ * Black Duck Software ("Confidential Information"). You shall not
+ * disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Black Duck Software.
+ */
+package com.blackducksoftware.integration.hub.detect.nameversion
+
+import com.blackducksoftware.integration.hub.bdio.simple.model.Forge
+
+interface NodeMetadata {
+    public Forge getForge()
+
+    public void setForge(Forge forge)
+}
