+++ src/test/groovy/com/blackducksoftware/integration/hub/detect/bomtool/yarn/YarnPackagerTest.groovy
@@ -17,11 +17,11 @@ import org.junit.Before
 import org.junit.Test
 
 import com.blackducksoftware.integration.hub.bdio.simple.model.DependencyNode
-import com.blackducksoftware.integration.hub.detect.nameversion.LinkMetadata
-import com.blackducksoftware.integration.hub.detect.nameversion.LinkedNameVersionNodeBuilder
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNode
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeImpl
 import com.blackducksoftware.integration.hub.detect.nameversion.NameVersionNodeTransformer
+import com.blackducksoftware.integration.hub.detect.nameversion.builder.LinkedNameVersionNodeBuilder
+import com.blackducksoftware.integration.hub.detect.nameversion.metadata.LinkMetadata
 import com.blackducksoftware.integration.hub.detect.testutils.TestUtil
 
 class YarnPackagerTest {
